+++ packages/splat-transform-native/source/include/node/api_splat.h
@@ -3,4 +3,5 @@
 
 namespace node_api::splat {
 Napi::Value generate_lod(const Napi::CallbackInfo& info);
+Napi::Value split(const Napi::CallbackInfo& info);
 } // namespace node_api::splat